requirements.txt
anyio==4.5.2
asgiref==3.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==4.2.11
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==21.2.0
idna==3.10
openai==1.51.0  # If using OpenAI API; remove if unnecessary
packaging==24.1
pillow==10.4.0
psycopg2==2.9.7
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.4.4  # Use an up-to-date version instead of pinned source
tzdata==2023.3  # Specify the latest release of tzdata if necessary
urllib3==2.2.3
whitenoise==6.6.0
cloudinary==1.37.0
django-cloudinary-storage==0.3.0
